feat: added a base heMteb task to reduce code duplication across all heMteb task
